blkmapd: remove pretty_sig()
This really only works against EMC servers.  There is at least one server
that returns unprintable signatures, which fill the log with garbage (the
spec does not mandate printable signatures).  It could be made more generic,
for example by checking each byte for isprint().  But the signatures are
really only of interest to developers, not admins, so it seems better to
just remove it.

Signed-off-by: Jim Rees <rees@umich.edu>
Signed-off-by: Steve Dickson <steved@redhat.com>