mountd: Set e_fslocdata field directly
To create an export entry for a junction, an options string is
constructed from the set of locations in the junction.  This options
string is then passed to mkexportent() where it is parsed and
converted into an exportent.

There is only one export option that is used to create a junction's
exportent: "refer=".  When that option is parsed, it's value is
simply copied to a fresh string and planted in the new export's
e_fslocdata field.

Let's avoid the option parsing and extra string copy.  Construct
a string for the new e_fslocdata field and plant it in the exportent
directly.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>